Rewrote documentation for all 'calculateEffective...' functions and removed references to Legion spells and abilities. Functionality is unchanged.
